Add health-poison scenario with fixed floor
In health gathering scenarios the floor is animated, which could lead to
a situation where two "screenshots" are almost the same except only for
the floor.

Even more: if we set a constant game seed, all medkits, poisons and the
player will be in the same positions at the start of the game, but the
floor can be in a different frame of animation. This will lead to two
theoretically equal states for us - because the floor is not important -,
but with different state values for the player - because the screen will
be different.

In practice, the agent will make different actions in games with the same
seed. In other words, the same environments become nondeterministic.

Changing the animated floor for a fixed one solves this problem.