Fixed poison sprite animation
The poison from health_poison_rewards_floor was also animated. So, the
same problem of the old animated floor applies: using the same game seed,
the initial (and the following) states could not be the same, if the
frame of the poison animation is different on them. This will imply in a
difference of only a few pixels. Although this difference could not be
significant, it still will imply in two different states for the agent.