Create Deep Q-Network agent
I created an agent based on Deep Q-Network to play Doom. It uses the
game screen as an input to the network to decide what action it should
make. More details in /pb/README.md.

The code was based in examples/python/learning_tensorflow.py but with a
different network setup.