Remove max pooling from DQN
We can say that pooling implies in an insensibility to features' spatial
position, that is, we lose information about locations of features. That
is not good in cases where the location is important, like a game.

Also, a minor change in code was made in setting learn/load the model.
Now, to load and/or train the agent you will need to set only two
booleans (learn_model and load_model) instead of the previous four.