Create a script to save frames of agent's vision
This code is intended to save every frame of a recorded play (.lmp files)
in order to create a video of the vision of the agent. As far as I know,
ViZDoom environment saves videos of only the raw screen buffer. However,
the agent receives a preprocessed image and I wanted to record a video of
its point of view. In order to do this, I needed to save every frame of a
play inside a folder and then use the function below to create the video.

I used this code to created the videos:
    $ ffmpeg -framerate 20 -i image_%00d.png -c:v libx264 -profile:v high -crf 20 -pix_fmt yuv420p output.mp4

I copied it from here (https://askubuntu.com/questions/610903/how-can-i-create-a-video-file-from-a-set-of-jpg-images)
and you can find more information about it there.

Note that code for saving frames is not optimized yet and certainly can
be improved. If you have some advice or wrote a better code, please share
with me and I will replace it.