Set the network in the agent
The network was configured inside his own file. Now, it is the agent that
should configure the network. This solves, among other things, a big
problem with dropout. The dropout probability was given only in the
constructor of the class, which means that the probability remained the
same. In practice, if we start the network to have dropout, we will have
it during training and testing.

Also, the max pooling was removed from the network.

Obs.: a dropout probability of 0.8 means that 80% of the network will
remain. To set dropout off, you have to set the probability to 1.0.