Load settings from config file
All settings were coded inside the code of the agents, thus to change a
single option it was needed to edit the source. Now, to change any option
one can simply edit the settings.cfg file, without modifying the agent
source code.

This improvement will allow an easy way of changing the setup of game and
model, training and testing regime, network architecture, and learning
parameters.