Fixed basic Deathmatch scenario
Previously, an agent who died during an episode respawned in the center
of the room, allowing a basic optimal behavior to kill the oponent and
going to the center of the room and shooting all the time.

Now the agent who dies during an episode respawns in a random position
along his axis.