Save frames of two players side by side
The scripts save_episodes and save_multiplayer only create the frames of
a single agent. However, to properly see the plays of two agents facing
themselves, it is good to have the frames of both agents side by side at
the same time. This new script was created to save exactly that.